Commit routes and models
committing work from yesterday, should have done at the end of day.
noticed ths morning that I was working in the main branch and need to
checkout a new branch - will be doing so.
Added 2 resources, piece and show, which currently have their own collections and
 have ownership established. the relationship between them is not fully
established but will be implemented in V2.0. I got ahead of myself.
    both user and show have an array field for pieces which i intend to
    fill with references and populate. I do not know if making pieces a
    subdoc of user was a better idea or not.
Routes have been written for both pice and show but are very v1.0.
the exception is for the show resource on a get by id, only requiring
authorization if it's isPrivate variable is = true. currently untested.

next steps are to checkout new branch, test crud routes for user and piece
with postman. Then focus will be on front end forms building the data for
those requests