Committing updated routes after testing
Have completed v1.0 and 1.5 level of crud on user and pieces, fixed typo in routes to include the 's' in the filename.

Tested user routes and added fields to the create to include the userName. Also added a update-profile route to allow the user to fill out their bio, mediums, and change their userName. NOTE: each field doesn't need to be populated, has 'if's. All are authorized routes

Piece routes all are functional, including the patch, which DOES require every field on the request currently. All require/generate ownership.